[TKT-15] refactor: 계정 등록 도메인 서비스와 애플리케이션 서비스 구분
