+++ src/main/java/com/trillion/tikitaka/domain/registration/domain/RegistrationDomainService.java
@@ -1,9 +1,7 @@
-package com.trillion.tikitaka.domain.registration.application;
+package com.trillion.tikitaka.domain.registration.domain;
 
 import org.springframework.stereotype.Service;
 
-import com.trillion.tikitaka.domain.registration.domain.Registration;
-import com.trillion.tikitaka.domain.registration.domain.RegistrationStatus;
 import com.trillion.tikitaka.global.exception.BusinessException;
 import com.trillion.tikitaka.global.exception.ErrorCode;
 
@@ -12,6 +10,14 @@ import lombok.RequiredArgsConstructor;
 @Service
 @RequiredArgsConstructor
 public class RegistrationDomainService {
+
+	public Registration createRegistration(String username, String email) {
+		return Registration.builder()
+			.username(username)
+			.email(email)
+			.status(RegistrationStatus.PENDING)
+			.build();
+	}
 
 	public String approveRegistration(Registration registration, String reason) {
 		if (registration.getStatus() != RegistrationStatus.PENDING) {
